# making kmeans take AbstractMatrix instead of Matrix