See #136 and #129. 

I duplicated the existing tests to take effectively the same matrix but now wrapped in a `Transpose` object.

Benchmarks on 1.1 / libopenlibm for what it's worth

**Before**

```julia
julia> seed!(1234)
julia> X = randn(5000, 200);
julia> @btime kmeans($X, 2);
3.138 ms (35 allocations: 95.34 KiB)
julia> @btime kmeans(transpose($Xt), 2);          # ERRORS
julia> @btime kmeans(copy(transpose($Xt)), 2);
4.835 ms (37 allocations: 7.72 MiB)
```

**After**
```julia
julia> seed!(1234)
julia> X = randn(5000, 200);
julia> @btime kmeans($X, 2);
3.158 ms (35 allocations: 95.34 KiB)
julia> @btime kmeans(transpose($Xt), 2);     # WORKS
21.913 ms (54 allocations: 96.10 KiB)
julia> @btime kmeans(copy(transpose($Xt)), 2);
  4.782 ms (37 allocations: 7.72 MiB)
```

So in short, the performances are identical, the only difference is that now it works if the user passes a transpose which I think is desirable. 
Of course we can argue that there is a performance drop if that's the case and it could be suggested that the user should do `copy(transpose(X))` if they can. My opinion is that were we should suggest it we should not force it and let the user decide whether that's a good idea for them or not. (the copy could be problematic if the data is very large).