@alyst 2019-01-25
Thanks! Besides `Transpose` there are other `AbstractMatrix` subtypes, which may work well without `collect()` (in particular, `SubArray`), so these changes would be quite useful.