@tlienart 2019-01-26
I made all the changes requested by @nalimilan and some of the changes requested by @alyst. 

Re @alyst: since in `kmeans!` there are now effectively those lines:

```julia
assignments = Vector{Int}(undef, n)
costs = Vector{Float64}(undef, n)
counts = Vector{Int}(undef, k)
cweights = Vector{Float64}(undef, k)
```

I don't think it's useful to have subsequent helper functions take `AbstractVector` etc. I tried to make changes that I thought would help towards generality though and, where appropriate, tried to be explicit about the type if the code warrants it. I hope the changes are ok. 

(on cosmetics, I hope you won't mind too much, I was just reviewing my changes and ended up doing a bit of cleaning up...)