@tlienart 2019-01-28
Thanks for the detailed feedback. I think I applied all the changes requested apart from 2 where I'm not sure I understood the comment and may have not interpreted it correctly.

Apologies for the bloated PR I wasn't expecting to have to touch so many parts of the code 🙈