@tlienart 2019-01-28
I've applied the changes in 16ae803. 

I must admit there are some things that I still don't like with the types (e.g. centers are updated with X*w, but X and c have the same type while w could potentially be something else) but this PR is already too messy so that would be for another issue.