@tlienart 2019-01-28
Apart from the broadcasting questions which are good for another PR (and possible refactoring of the code), is there anything else left to change before a merge?