@alyst 2019-01-28
LGTM except `weights` declaration in `kmeans!()`.