@tlienart 2019-01-29
> Fine with me, but then please revert all broadcasting changes which aren't related to the main purpose of this PR. It's never a good idea to mix different issues in a single PR (as we can see).

I don't think there's any non-trivial change in that regards that are left?